backend/requirements.txt
Django==1.8.4
Pillow==2.9.0
Twisted==15.4.0
amqp==1.4.6
anyjson==0.3.3
appnope==0.1.0
billiard==3.3.0.20
celery==3.1.18
cffi==1.2.1
characteristic==14.3.0
coverage==3.7.1
cryptography==1.0.1
cssselect==0.9.1
decorator==4.0.2
django-choices==1.4.2
django-coverage==1.2.4
django-markup==0.4
djangorestframework==3.2.3
drf-nested-routers==0.10.0
enum34==1.0.4
flake8==2.4.1
funcsigs==0.4
gnureadline==6.3.3
idna==2.0
ipaddress==1.0.14
ipython==4.0.0
ipython-genutils==0.1.0
kombu==3.0.26
lxml==3.4.4
mccabe==0.3.1
mock==1.3.0
path.py==8.1.1
pbr==1.8.0
pep8==1.5.7
pexpect==3.3
pickleshare==0.5
psycopg2==2.6.1
pyOpenSSL==0.15.1
pyasn1==0.1.8
pyasn1-modules==0.0.7
pycparser==2.14
pyflakes==0.8.1
pytz==2015.4
queuelib==1.4.2
redis==2.10.3
requests==2.7.0
service-identity==14.0.0
simplegeneric==0.8.1
six==1.9.0
traitlets==4.0.0
urllib3==1.12
w3lib==1.12.0
wsgiref==0.1.2
zope.interface==4.1.2
